consolidate decompression tests
